# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Inventario",
        "description": "Documentación de la API del sistema de inventario",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/negocios": {
            "get": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Obtener negocios del usuario autenticado",
                "operationId": "6356ca76f5729642d8bffcb57a7e6ddf",
                "responses": {
                    "200": {
                        "description": "Lista de negocios"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Crear un nuevo negocio",
                "operationId": "a6c057c5d43494657c978c845496f57a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Tienda Central"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Principal 123"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "70000000"
                                    },
                                    "logo": {
                                        "type": "string",
                                        "example": "logo.png"
                                    },
                                    "estado": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Negocio creado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/negocios/{id}": {
            "get": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Obtener un negocio por ID",
                "operationId": "c8a06abb16047eed78b3e1977f0da785",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del negocio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Negocio encontrado"
                    },
                    "404": {
                        "description": "Negocio no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Actualizar negocio",
                "operationId": "dabd2bebb9ad7661d4de8fd489456238",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del negocio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Negocio actualizado"
                    },
                    "404": {
                        "description": "Negocio no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Negocios"
                ],
                "summary": "Eliminar negocio",
                "operationId": "4884926613d3ffe371d75cc07723dbcc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del negocio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Negocio eliminado"
                    },
                    "404": {
                        "description": "Negocio no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}